requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
fiona==1.9.5
fonttools==4.44.0
geopandas==0.14.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pycountry==22.3.5
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
